# docs/api/openapi-v2.yaml
openapi: 3.0.3
info:
  title: Poco.ai V2 Policy Analysis API
  description: |
    Country-agnostic insurance policy analysis and comparison platform with advanced PII protection and real-time processing.
    
    ## Features
    - Multi-country support (AU, SG, NZ) with flexible regulatory frameworks
    - Real-time analysis progress via Server-Sent Events
    - Advanced PII detection, encryption, and anonymization
    - AI-powered policy feature extraction and comparison
    - Comprehensive provider recommendations with detailed scoring
    
    ## Authentication
    Currently using session-based authentication. API key authentication coming in future releases.
    
    ## Rate Limiting
    - Analysis requests: 10 per hour per IP
    - Progress requests: 60 per minute per session
    - Results requests: 30 per minute per session
  version: 2.0.0
  contact:
    name: Poco.ai API Support
    url: https://poco.ai/support
    email: api-support@poco.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://poco.ai/terms
servers:
  - url: https://poco.ai/api/v2
    description: Production server
  - url: https://staging.poco.ai/api/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
paths:
  /v2/{country}/policies/analyze:
    post:
      summary: Start Policy Analysis
      description: |
        Initiates comprehensive analysis of an insurance policy document for the specified country.
        The analysis includes PII detection/protection, AI feature extraction, and provider comparison.
      operationId: analyzePolicy
      tags:
        - Policy Analysis
      parameters:
        - name: country
          in: path
          required: true
          description: ISO 3166-1 alpha-2 country code
          schema:
            type: string
            enum: [AU, SG, NZ]
          examples:
            australia:
              value: AU
              description: Australia
            singapore:
              value: SG
              description: Singapore
            new_zealand:
              value: NZ
              description: New Zealand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V2AnalysisRequest'
            examples:
              basic_analysis:
                summary: Basic policy analysis
                value:
                  policy_text: "Hospital Insurance Policy\nProvider: HCF\nPremium: $250/month\nCoverage: Silver Hospital Cover\nExcess: $500\n\nCovered Services:\n- Emergency surgery\n- Cardiology\n- Orthopedics\n\nExclusions:\n- Cosmetic surgery\n- Dental (except emergency)"
              detailed_analysis:
                summary: Analysis with preferences and options
                value:
                  policy_text: "PRIVATE HEALTH INSURANCE STATEMENT\nPolicyowner: John Smith\nPolicy Number: PHI123456\nPremium: $280.50 monthly\nCoverage Tier: Gold Hospital + Extras\nExcess: $500 hospital, $100 extras\n\nHospital Benefits:\n- Private room guarantee\n- No waiting periods for emergencies\n- Surgery with choice of doctor\n- Cardiac procedures\n- Cancer treatment\n\nExtras Benefits:\n- Dental: 80% up to $1,200\n- Optical: $300 annual limit\n- Physiotherapy: 80% up to $800\n- Psychology: $150 per session, 10 sessions"
                  user_preferences:
                    max_premium: 350
                    preferred_providers: ["MEDIBANK", "BUPA", "NIB"]
                    must_have_features: ["cardiac_procedures", "cancer_treatment"]
                    comparison_count: 5
                  analysis_options:
                    include_cost_comparison: true
                    include_feature_gap_analysis: true
                    include_provider_recommendations: true
                    priority_weights:
                      cost: 0.3
                      coverage: 0.5
                      provider_reputation: 0.1
                      customer_service: 0.1
      responses:
        '202':
          description: Analysis started successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per hour
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStartResponse'
              examples:
                success:
                  value:
                    session_id: "sess_1732276800_abc123def456"
                    status: "processing"
                    message: "Analysis started successfully"
                    progress_url: "/api/v2/au/policies/progress/sess_1732276800_abc123def456"
                    results_url: "/api/v2/au/policies/results/sess_1732276800_abc123def456"
                    estimated_completion: "2025-11-22T10:03:00Z"
                    country_config:
                      code: "AU"
                      name: "Australia"
                      currency: "AUD"
                      supported_providers: ["HCF", "MEDIBANK", "BUPA", "NIB", "AHSA"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v2/{country}/policies/progress/{sessionId}:
    get:
      summary: Get Analysis Progress
      description: |
        Retrieves real-time progress information for an ongoing policy analysis.
        Use this endpoint to poll for progress updates or implement progress bars.
      operationId: getAnalysisProgress
      tags:
        - Policy Analysis
      parameters:
        - name: country
          in: path
          required: true
          description: ISO 3166-1 alpha-2 country code
          schema:
            type: string
            enum: [AU, SG, NZ]
        - name: sessionId
          in: path
          required: true
          description: Session ID returned from analysis request
          schema:
            type: string
            pattern: '^sess_[0-9a-f]{8,64}$'
            example: "sess_1732276800_abc123def456"
      responses:
        '200':
          description: Progress information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2SessionProgress'
              examples:
                in_progress:
                  value:
                    session_id: "sess_1732276800_abc123def456"
                    stage: "ai_analysis"
                    progress: 45
                    message: "Analyzing policy features with Gemini 1.5 Pro"
                    stages:
                      - name: "document_upload"
                        status: "completed"
                        started_at: "2025-11-22T10:01:00Z"
                        completed_at: "2025-11-22T10:01:05Z"
                      - name: "pii_protection"
                        status: "completed"
                        started_at: "2025-11-22T10:01:05Z"
                        completed_at: "2025-11-22T10:01:15Z"
                      - name: "ai_analysis"
                        status: "in_progress"
                        started_at: "2025-11-22T10:01:15Z"
                      - name: "policy_comparison"
                        status: "pending"
                      - name: "recommendation_ranking"
                        status: "pending"
                    estimated_time_remaining: 75
                completed:
                  value:
                    session_id: "sess_1732276800_abc123def456"
                    stage: "completed"
                    progress: 100
                    message: "Analysis completed successfully"
                    stages:
                      - name: "document_upload"
                        status: "completed"
                        started_at: "2025-11-22T10:01:00Z"
                        completed_at: "2025-11-22T10:01:05Z"
                      - name: "pii_protection"
                        status: "completed"
                        started_at: "2025-11-22T10:01:05Z"
                        completed_at: "2025-11-22T10:01:15Z"
                      - name: "ai_analysis"
                        status: "completed"
                        started_at: "2025-11-22T10:01:15Z"
                        completed_at: "2025-11-22T10:01:45Z"
                      - name: "policy_comparison"
                        status: "completed"
                        started_at: "2025-11-22T10:01:45Z"
                        completed_at: "2025-11-22T10:02:30Z"
                      - name: "recommendation_ranking"
                        status: "completed"
                        started_at: "2025-11-22T10:02:30Z"
                        completed_at: "2025-11-22T10:03:00Z"
                    estimated_time_remaining: 0
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v2/{country}/policies/results/{sessionId}:
    get:
      summary: Get Analysis Results
      description: |
        Retrieves complete analysis results including policy comparison, recommendations, and cost analysis.
        Results are available for 7 days after completion.
      operationId: getAnalysisResults
      tags:
        - Policy Analysis
      parameters:
        - name: country
          in: path
          required: true
          description: ISO 3166-1 alpha-2 country code
          schema:
            type: string
            enum: [AU, SG, NZ]
        - name: sessionId
          in: path
          required: true
          description: Session ID from completed analysis
          schema:
            type: string
            pattern: '^sess_[0-9a-f]{8,64}$'
        - name: include_raw_data
          in: query
          required: false
          description: Include raw analysis data in response
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          required: false
          description: Response format
          schema:
            type: string
            enum: [json, pdf]
            default: json
      responses:
        '200':
          description: Analysis results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2AnalysisResponse'
              examples:
                complete_results:
                  $ref: '#/components/examples/CompleteAnalysisResults'
            application/pdf:
              schema:
                type: string
                format: binary
              description: PDF report of analysis results
        '202':
          description: Analysis still in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatus'
              examples:
                still_processing:
                  value:
                    status: "processing"
                    session_id: "sess_1732276800_abc123def456"
                    message: "Analysis still in progress"
                    progress: 75
                    estimated_completion: "2025-11-22T10:04:30Z"
                    progress_url: "/api/v2/au/policies/progress/sess_1732276800_abc123def456"
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '410':
          $ref: '#/components/responses/ResultsExpired'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    V2AnalysisRequest:
      type: object
      required:
        - policy_text
      properties:
        policy_text:
          type: string
          minLength: 100
          maxLength: 50000
          description: Raw text content of the insurance policy document
          example: "Hospital Insurance Policy\nProvider: HCF\nPremium: $250/month..."
        user_preferences:
          type: object
          description: User preferences for analysis and recommendations
          properties:
            max_premium:
              type: number
              minimum: 0
              maximum: 10000
              description: Maximum acceptable monthly premium
              example: 300
            preferred_providers:
              type: array
              items:
                type: string
                enum: [HCF, MEDIBANK, BUPA, NIB, AHSA, AHM, TEACHERS_HEALTH]
              maxItems: 10
              description: List of preferred insurance providers
              example: ["MEDIBANK", "BUPA"]
            must_have_features:
              type: array
              items:
                type: string
              maxItems: 20
              description: Required coverage features
              example: ["emergency_surgery", "cardiac_procedures"]
            comparison_count:
              type: integer
              minimum: 1
              maximum: 10
              default: 5
              description: Number of policy recommendations to return
              example: 3
        analysis_options:
          type: object
          description: Options to customize the analysis process
          properties:
            include_cost_comparison:
              type: boolean
              default: true
              description: Include detailed cost comparison analysis
            include_feature_gap_analysis:
              type: boolean
              default: true
              description: Include feature gap analysis
            include_provider_recommendations:
              type: boolean
              default: true
              description: Include provider-specific recommendations
            priority_weights:
              type: object
              description: Weighting for different factors in recommendations
              properties:
                cost:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  description: Weight for cost considerations
                  example: 0.4
                coverage:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  description: Weight for coverage quality
                  example: 0.4
                provider_reputation:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  description: Weight for provider reputation
                  example: 0.1
                customer_service:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  description: Weight for customer service quality
                  example: 0.1
...
    V2SessionProgress:
      type: object
      required:
        - session_id
        - stage
        - progress
        - message
        - stages
      properties:
        session_id:
          type: string
          pattern: '^sess_[0-9a-f]{8,64}$'
          description: Session identifier
          example: "sess_1732276800_abc123def456"
        stage:
          type: string
          enum: [document_upload, pii_protection, ai_analysis, policy_comparison, recommendation_ranking, completed, error]
          description: Current processing stage
          example: "ai_analysis"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall progress percentage
          example: 45
        message:
          type: string
          maxLength: 200
          description: Human-readable progress message
          example: "Analyzing policy features with Gemini 1.5 Pro"
        stages:
          type: array
          items:
            $ref: '#/components/schemas/ProcessingStage'
          description: Detailed stage information
        estimated_time_remaining:
          type: integer
          minimum: 0
          description: Estimated seconds until completion
          example: 75
        error:
          type: string
          description: Error message if processing failed
...
    AnalysisStartResponse:
      type: object
      required:
        - session_id
        - status
        - message
        - progress_url
        - results_url
        - estimated_completion
        - country_config
      properties:
        session_id:
          type: string
          pattern: '^sess_[0-9a-f]{8,64}$'
          description: Unique session identifier
          example: "sess_1732276800_abc123def456"
        status:
          type: string
          enum: [processing]
          description: Analysis status
          example: "processing"
        message:
          type: string
          description: Success message
          example: "Analysis started successfully"
        progress_url:
          type: string
          format: uri
          description: URL to check analysis progress
          example: "/api/v2/au/policies/progress/sess_1732276800_abc123def456"
        results_url:
          type: string
          format: uri
          description: URL to retrieve results when complete
          example: "/api/v2/au/policies/results/sess_1732276800_abc123def456"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-11-22T10:03:00Z"
        country_config:
          type: object
          required:
            - code
            - name
            - currency
            - supported_providers
          properties:
            code:
              type: string
              enum: [AU, SG, NZ]
              description: Country code
              example: "AU"
            name:
              type: string
              description: Country name
              example: "Australia"
            currency:
              type: string
              enum: [AUD, SGD, NZD]
              description: Country currency
              example: "AUD"
            supported_providers:
              type: array
              items:
                type: string
              description: List of supported insurance providers
              example: ["HCF", "MEDIBANK", "BUPA", "NIB", "AHSA"]
...
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            invalid_policy_text:
              value:
                error:
                  code: "INVALID_POLICY_TEXT"
                  message: "Policy text is required and must be at least 100 characters"
                  details: "Provided text was only 45 characters long"
                  suggestion: "Please provide the complete policy document text"
                timestamp: "2025-11-22T10:01:00Z"
                country_code: "AU"
            invalid_country_code:
              value:
                error:
                  code: "INVALID_COUNTRY_CODE"
                  message: "Country code must be 2-letter ISO code"
                  details: "Received 'AUS' but expected 'AU'"
                  suggestion: "Use 2-letter country codes: AU, SG, NZ"
                timestamp: "2025-11-22T10:01:00Z"
    UnprocessableEntity:
      description: Unprocessable entity - Business logic validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            unsupported_country:
              value:
                error:
                  code: "UNSUPPORTED_COUNTRY"
                  message: "Country 'US' is not supported"
                  details: "Supported countries: AU, SG, NZ"
                  suggestion: "Use one of the supported country codes"
                timestamp: "2025-11-22T10:01:00Z"
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until next request allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            rate_limit_exceeded:
              value:
                error:
                  code: "RATE_LIMIT_EXCEEDED"
                  message: "Too many requests. Please try again in 60 seconds"
                  details: "Limit: 10 requests per hour per IP"
                  suggestion: "Wait for the rate limit to reset or upgrade to a higher tier"
                timestamp: "2025-11-22T10:01:00Z"
    SessionNotFound:
      description: Session not found or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            session_not_found:
              value:
                error:
                  code: "SESSION_NOT_FOUND"
                  message: "Session 'sess_invalid123' not found or expired"
                  details: "Sessions expire after 7 days of inactivity"
                  suggestion: "Check the session ID or start a new analysis"
                timestamp: "2025-11-22T10:01:00Z"
                country_code: "AU"
    ResultsExpired:
      description: Analysis results have expired and been purged
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            results_expired:
              value:
                error:
                  code: "RESULTS_EXPIRED"
                  message: "Analysis results have expired and been purged"
                  details: "Results are available for 7 days after completion"
                  suggestion: "Please run a new analysis"
                timestamp: "2025-11-22T10:01:00Z"
                country_code: "AU"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V2ErrorResponse'
          examples:
            ai_service_error:
              value:
                error:
                  code: "AI_SERVICE_UNAVAILABLE"
                  message: "AI analysis service is currently unavailable"
                  details: "Gemini API is experiencing high load"
                  suggestion: "Please try again in a few minutes"
                timestamp: "2025-11-22T10:01:00Z"
            pii_protection_error:
              value:
                error:
                  code: "PII_PROTECTION_ERROR"
                  message: "Privacy processing failed"
                  details: "Unable to encrypt sensitive information"
                  suggestion: "Please contact support if this error persists"
                timestamp: "2025-11-22T10:01:00Z"
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Session-based authentication (current implementation)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (coming soon)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (future implementation)
security:
  - SessionAuth: []
...